backend/requirements.txt
# Core FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI and ML
openai==1.3.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
torch==2.1.2
transformers==4.36.2

# Web and API
requests==2.31.0
beautifulsoup4==4.13.3
python-jose==3.3.0
passlib==1.7.4

# File Processing
python-magic==0.4.27
pdfminer.six==20201018

# Database
SQLAlchemy==2.0.38

# Utilities
pydantic==2.10.6
typing-extensions==4.12.2